refractor: move indexMapping to component
