+++ public/pages/CreateIndex/components/AliasSelect/index.tsx
@@ -3,15 +3,16 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useCallback, useContext, useState } from "react";
+import React, { useCallback, useState } from "react";
 import { EuiComboBox } from "@elastic/eui";
 import { useEffect } from "react";
-import { ServicesContext } from "../../../../services";
+import { debounce } from "lodash";
 import { ServerResponse } from "../../../../../server/models/types";
 
 export interface AliasSelectProps {
   value?: Record<string, {}>;
   onChange: (value: AliasSelectProps["value"]) => void;
+  refreshOptions: (aliasName: string) => Promise<ServerResponse<{ alias: string }[]>>;
 }
 
 const transformObjToArray = (obj: AliasSelectProps["value"]): { label: string }[] => {
@@ -21,36 +22,31 @@ const transformArrayToObj = (array: { label: string }[]): AliasSelectProps["valu
   return array.reduce((total, current) => ({ ...total, [current.label]: {} }), {});
 };
 
-const AliasSelect = ({ value, onChange }: AliasSelectProps) => {
+const AliasSelect = (props: AliasSelectProps) => {
+  const { value, onChange } = props;
   const finalValue = transformObjToArray(value);
-  const services = useContext(ServicesContext);
   const [allOptions, setAllOptions] = useState([] as { label: string }[]);
   const [isLoading, setIsLoading] = useState(false);
-  const refreshOptions = useCallback(({ aliasName }) => {
-    setIsLoading(true);
-    services?.commonService
-      .apiCaller({
-        endpoint: "cat.aliases",
-        method: "GET",
-        data: {
-          format: "json",
-          name: aliasName,
-          expand_wildcards: "open",
-        },
-      })
-      .then((res: ServerResponse<{ alias: string }[]>) => {
-        if (res.ok && res.response) {
-          setAllOptions(
-            [...new Set(res.response.map((item) => item.alias).filter((item) => !item.startsWith(".")))].map((item) => ({ label: item }))
-          );
-        } else {
-          setAllOptions([]);
-        }
-      })
-      .finally(() => {
-        setIsLoading(false);
-      });
-  }, []);
+  const refreshOptions = useCallback(
+    debounce(({ aliasName }) => {
+      setIsLoading(true);
+      props
+        .refreshOptions(aliasName)
+        .then((res: ServerResponse<{ alias: string }[]>) => {
+          if (res.ok && res.response) {
+            setAllOptions(
+              [...new Set(res.response.map((item) => item.alias).filter((item) => !item.startsWith(".")))].map((item) => ({ label: item }))
+            );
+          } else {
+            setAllOptions([]);
+          }
+        })
+        .finally(() => {
+          setIsLoading(false);
+        });
+    }, 500),
+    []
+  );
   useEffect(() => {
     refreshOptions({});
   }, []);
@@ -67,7 +63,6 @@ const AliasSelect = ({ value, onChange }: AliasSelectProps) => {
 
     // Create the option if it doesn't exist.
     if (flattenedOptions.findIndex((option: { label: string }) => option.label.trim().toLowerCase() === normalizedSearchValue) === -1) {
-      // Simulate creating this option on the server.
       setAllOptions([...allOptions, newOption]);
       onChange(transformArrayToObj([...finalValue, newOption]));
     }
@@ -81,6 +76,7 @@ const AliasSelect = ({ value, onChange }: AliasSelectProps) => {
       options={allOptions}
       isLoading={isLoading}
       onSearchChange={(searchValue) => {
+        setIsLoading(true);
         refreshOptions({ aliasName: searchValue });
       }}
       customOptionText={"Add {searchValue} as a new alias"}
