--- public/pages/CreateIndex/containers/AliasSelect/index.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-import AliasSelect from "./AliasSelect";
-
-export default AliasSelect;
